# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golf caddie backend — players, rounds, shots and course data.",
        "title": "Agentic Caddie API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/holes": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holes"
                ],
                "summary": "Record a hole in a round",
                "parameters": [
                    {
                        "description": "Hole to record",
                        "name": "hole",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createHoleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.holeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/holes/round/{roundId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holes"
                ],
                "summary": "List holes for a round",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Round ID",
                        "name": "roundId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.holeResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid round ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/holes/round/{roundId}/number/{holeNumber}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holes"
                ],
                "summary": "Get a hole by round and hole number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Round ID",
                        "name": "roundId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Hole number",
                        "name": "holeNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.holeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Hole not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/holes/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holes"
                ],
                "summary": "Get a hole by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hole ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.holeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Hole not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "holes"
                ],
                "summary": "Update a hole's scorecard data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hole ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated fields",
                        "name": "hole",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateHoleRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "holes"
                ],
                "summary": "Delete a hole record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hole ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.createHoleRequest": {
            "type": "object",
            "properties": {
                "course_hole_id": {
                    "type": "integer",
                    "example": 1
                },
                "flag_position": {
                    "type": "string",
                    "example": "middle_centre"
                },
                "gir": {
                    "type": "boolean",
                    "example": true
                },
                "hole_number": {
                    "type": "integer",
                    "example": 1
                },
                "penalty": {
                    "type": "boolean",
                    "example": false
                },
                "points": {
                    "type": "integer",
                    "example": 2
                },
                "putts": {
                    "type": "integer",
                    "example": 2
                },
                "round_id": {
                    "type": "integer",
                    "example": 1
                },
                "score": {
                    "type": "integer",
                    "example": 4
                },
                "scramble_save": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.holeResponse": {
            "type": "object",
            "properties": {
                "course_hole_id": {
                    "type": "integer"
                },
                "flag_position": {
                    "type": "string"
                },
                "gir": {
                    "type": "boolean"
                },
                "hole_number": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "penalty": {
                    "type": "boolean"
                },
                "points": {
                    "type": "integer"
                },
                "putts": {
                    "type": "integer"
                },
                "round_id": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "scramble_save": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.updateHoleRequest": {
            "type": "object",
            "properties": {
                "flag_position": {
                    "type": "string",
                    "example": "middle_centre"
                },
                "gir": {
                    "type": "boolean",
                    "example": true
                },
                "penalty": {
                    "type": "boolean",
                    "example": false
                },
                "points": {
                    "type": "integer",
                    "example": 2
                },
                "putts": {
                    "type": "integer",
                    "example": 2
                },
                "score": {
                    "type": "integer",
                    "example": 4
                },
                "scramble_save": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    }
}